Feature Extraction Input Dimensions
Feature extraction input was passed in as image format, which worked but was inconsistent with the expected format of (windows x channels x samples).

Modified feature extraction to occur before remapping to spatial image format. This gives the same results as before, but feature extraction is ensured to be consistent with what is expected from the libemg.feature_extractor.FeatureExtractor class.