Third DOF Display Parameter
The way that the third DOF was displayed was determined internally based on the range of the third DOF values. This was kind of unintuitive and was inconsistent with the other DOFs.

Modified this so you can pass in a third_dof_display parameter that determines the method for modifying the third DOF. This way the same [-1, 1] range is passed in for all DOFs and is mapped to degrees if they want rotation.