Plot Icon Abstract Method
make_regression_training_gif() had a lot of hidden functions for plotting and the plotting function used was decided by specific parameters.

Refactoring this so the RegressionAnimator must implement a plot_icon() method that will handle this. This way the user can create their own custom plotting targets as well.