Add time per distance parameter to PlotAnimator
PlotAnimator had a default value of 2 seconds for a point to travel a
distance of 1. This made it so the user couldn't modify it if they
wanted a different amount of time.

Added a time per distance (tpd) parameter to allow for customization of
time to travel a unit distance.