Training Deep Regression Models (#32)
* Add model fitting for deep regression models

A fit method was not implemented for deep regression models.

Added a fit method very similar to that of EMGClassifier.

* Create EMGPredictor class definition

EMGClassifier and EMGRegressor contained a lot of common code. This
meant that changes / fixes in one had to be modified separately for both
implementations.

Added a base class that both can inherit from.

* EMGClassifier and EMGRegressor inherit from EMGPredictor

Added inheritance for EMGClassifier and EMGRegressor. Removed redundant
methods that are now in EMGPredictor.

* Move model to constructor

Previous implementation had the model passed in as a parameter to fit().
This was inconsistent with libraries like scikit-learn.

Moved the model to the constructor instead. Moved model validation to
EMGPredictor so both children could use it.

* Moved feature parameter functions to parent

Feature parameter functions could be used in both children, but they
were only found in the EMGClassifier.

Moved this functionality to the parent class.

* Create EMGRegressor constructor

* Add default value for model_parameters in constructors

* Remove unused parameter when fitting statistical models

* Update docstrings

* Add future ideas for EMGRegressor

* Rename emg_classifier.py

Renamed to emg_predictor.py because it doesn't only contain classifiers now.

* Implement prediction methods

Implemented prediction methods that give error messages if an invalid
model is passed in.

* Update package name in __init__.py

* Fix fit() docstring

* Add NotImplementedErrors to unimplemented methods

Methods just contained a noop line if they weren't implemented yet.

Modified these so they throw an error. This way the user doesn't have to
be confused that something didn't happen when they called the method.

* Fix docstrings

* Fix unpacking deep parameters when None are passed

The default value of fit_parameters when fitting the model was None.
When unpacking this value in _fit_deeplearning_model() an error was
raised because None can't be unpacked.

Added a check so variable is converted to an empty dictionary if None is
passed.

* Change references from emg_classifier to emg_predictor

Changed all references, including in examples and docs, of
emg_classifier to emg_predictor.