Collect regression labels in OfflineDataHandler.get_data()
Adding regression labels was slightly tedious because you had to call an
extra method (add_regression_labels()). This wasn't too bad for single
movements, but this was very annoying for reps that had different labels files because you'd have to loop through each type of label file and add OfflineDataHandlers together.

Added some of the logic from add_regression_labels to get_data(). The user can pass in a labels_filename and the labels will be added appropriately. Additionally, a single 'labels' field is added instead of separate ones for each DOF.