Extract data formatting to EMGPredictor
Formatting of data was different for EMGClassifier, EMGRegressor,
running, and fitting.

Extracted this process to a single method that formats data to an array
if needed and handles NaNs (if desired). Parameters for whether the
model should fix NaNs (+ if it should do this silently) was moved to the
constructor so it's consistent for fitting and running.