MetadataFetcher class
Removed the FileFetcher class because I think it's fine to expect users
to always use regex to filter out files.

Instead of explicitly having a separate parameter in get_data for
FilePackagers, added a MetadataFetcher class that the method will take
in. Now the RegexFilter will always filter the files and grab the
metadata, and they can pass in extra objects to grab other specific
metadata. Having this linked to something that grabbed a file didn't
make sense because they may need to grab multiple types of metadata from
a single file.