FilePackager consistency fixes
The align_method parameter only passed the packaged file data to the
callable instead of passing in the EMG data as well. Also, passing a
value to the load parameter meant the metadata would be returned before
aligning it, which was inconsistent with the other methods and meant the
user would have to implement this themselves.

Modified call to align_method to pass in the appropriate parameters and
stored result of the load method in a variable that is used for the rest
of data processing.