Regressor Convenience Methods (#39)
* Fix .fit() docstring

The `model` parameter was still in the docstring for .fit() even though
this method doesn't use that parameter anymore.

* Handle passing ndarray to EMGPredictor.fit()

EMGPredictor.fit() expected the training_features key to be a
dictionary. If an array was passed in then it would loop through each
row and stack them horizontally.

Added a check in EMGPredictor._format_data() that will just return the
array when passed in.

* Return predictions

predict() and predict_proba() were implemented in EMGPredictor, but
their values were not returned.

* EMGRegressor constructor documentation fixes

Modified inclusion of 'classifier' in regressor constructor docs and
added explanation of SVM.

* Collect regression labels in OfflineDataHandler.get_data()

Adding regression labels was slightly tedious because you had to call an
extra method (add_regression_labels()). This wasn't too bad for single
movements, but this was very annoying for reps that had different labels files because you'd have to loop through each type of label file and add OfflineDataHandlers together.

Added some of the logic from add_regression_labels to get_data(). The user can pass in a labels_filename and the labels will be added appropriately. Additionally, a single 'labels' field is added instead of separate ones for each DOF.

* Implement deadband utility for EMGRegressor

Added the ability to add a deadband around EMGRegressor predictions both
in the constructor and using a method.

* Fix deadband utility for EMGRegressor

Previous implementation modified the .run() method, which only works
when .run() is called but not .predict().

Overrode the .predict() method so now all predictions will be clipped
based on the deadband.

* Move deadband thresholding back to .run()

I had moved the deadband checking to the .predict() method, but this is
inconsistent with EMGClassifier. The .predict() method also isn't what
should be used, and they should really use .run() instead.

Therefore, I moved the deadband logic back to the .run() method.

* Rename EMGPredictor prediction methods

.predict() and .predict_proba() were not named to suggest that they
shouldn't be used, which was confusing because EMGClassifier and
EMGRegressor encourage use of the run() method.

Added underscores to tell the user that these methods shouldn't be used.

* Extract data formatting to EMGPredictor

Formatting of data was different for EMGClassifier, EMGRegressor,
running, and fitting.

Extracted this process to a single method that formats data to an array
if needed and handles NaNs (if desired). Parameters for whether the
model should fix NaNs (+ if it should do this silently) was moved to the
constructor so it's consistent for fitting and running.

* EMGRegressor visualize() method

EMGRegressor didn't have a visualize() method, making it have less
functionality than EMGClassifier.

Implemented a visualize() method that will plot the decision stream.

* EMGRegressor visualize docstring

Added docstring for EMGRegressor.visualize().

* Add analyze_regerssor() skeleton

OnlineEMGRegressor doesn't have a method equivalent to
OnlineEMGClassifier.analyze_classifier().

Added a skeleton that will be implemented later.