Remove file_data from RegexFilter.get_metadata
Requiring file_data to be passed in made the RegexFilter less useful for
other use cases (such as just grabbing from a filename outside
DataHandler).

Removed this parameter so only the filename is needed now and the
metadata is cast to an array in get_data() instead. Also moved the check
for 2D arrays into get_data() from ColumnFetcher so any MetadataFetcher
will be cast to 2D if needed.