Fix reassigning file_data bug
Last commit I changed a conditional statement so file_data was reassigned based on whether or not the EMG data was being grabbed from a column. This introduced a bug because the metadata fetcher used the file_data to grab other metadata, which didn't work when file_data was reassigned to only EMG data.

Reverted conditional statement to how it was previously.