Sort EMaGer columns based on channel index
Data streamed from EMaGer returned an array where the index of the
column didn't correspond to the channel index. This meant that a special
remap function was needed that reshaped and remapped columns.

Modified streamer so columns are now reordered per sample, removing the
need for a special remap function and avoiding confusion.