Handle single values from MetadataFetcher
MetadatFetcher stated in the docs that it expected an N x M array, but didn't throw an error as long as an array was returned.

Added a check to ensure that single values aren't being returned. If a single value array is returned, it is cast to an N x 1 array.