Convert Manager to SharedMemory
Previous implementation used a multiprocessing.Value and a string array, which is different from other low level memory interfaces in libemg (e.g., streamers) and isn't as fast.

Replaced with straight shared memory. This way it uses the same interface as the streamers, has a fixed buffer size, and is quicker.