Extract Fitts parameters to config dataclass
Managing contructor parameters and docstrings for Fitts and ISOFitts was quite tedious and would only get worse when adding more Fitts environments.

Extracted many configuration parameters (e.g., target size and color) to a dataclass. Now the docstring and parameters of the dataclass can be changed to update both environments.