Make OnlineDataHandler processes daemons
Processes are created for visualize and log_to_file, but they weren't created as daemon processes. This meant that users had to manually call stop_log() and stop_visualize(). Otherwise, the script would hang because these processes would continue even after the main script was finished.

Created these as daemons so they will be killed when the parent process is killed.